DjangoBackend/requirements.txt
Django==1.8.19
django-cors-headers==2.4.0
djangorestframework==3.4.0
MySQL-python==1.2.5
